fix hw03
