+++ server.js
@@ -1,12 +1,12 @@
 require("dotenv").config();
-const {app }= require("./src/app");
+const { app }= require("./src/app");
 const { connectMongo } = require("./src/db/connection");
 
 const PORT = process.env.PORT || 3000;
 
-async function start() {
+(async function () {
   try {
-   connectMongo();
+    await connectMongo();
     
     app.listen(PORT, () => {
       console.log(`Server is listening on port ${PORT}`);
@@ -15,6 +15,4 @@ async function start() {
   } catch (error) {
     console.error("Error at server launch", error.message);
   }
- }
- 
-start();
+})();
